requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
openai
wikipedia
python-dotenv
redis
typing-extensions

